requirements.txt
pythons=3.9
black
isort
flake8
pytorch
pre-commit
k3d
matplotlib